+++ src/Events/QueryAudited.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Teoprayoga\Teorion\Events;
+
+class QueryAudited
+{
+    public function __construct(public readonly array $record)
+    {
+    }
+}